re organize tests according to middleware and decorator scope
