+++ app/tests/test_usagesmiddleware.py
@@ -1,11 +1,12 @@
 from datetime import datetime
+import os
 import time
 
 import pytest
 
 from app.schemas.models import ModelType
 from app.sql.models import Usage
-from app.utils.variables import ENDPOINT__EMBEDDINGS, ENDPOINT__MODELS
+from app.utils.variables import ENDPOINT__EMBEDDINGS, ENDPOINT__MODELS, ENDPOINT__RERANK
 
 
 @pytest.mark.usefixtures("client")
@@ -34,15 +35,78 @@ class TestUsagesMiddleware:
         assert log.user is not None
         assert log.duration is not None
 
-    def test_log_with_invalid_json(self, client, db_session):
-        """Test handling of invalid JSON in request body"""
+    def test_audio(self, client, db_session):
+        """Test logging of an audio transcription request using log_usage decorator."""
+        from app.utils.variables import ENDPOINT__AUDIO_TRANSCRIPTIONS
 
-        previous = db_session.query(Usage).count()
-        invalid_json = "{"  # Invalid JSON
-        response = client.post(f"/v1{ENDPOINT__EMBEDDINGS}", data=invalid_json, headers={"Content-Type": "application/json"})
-        assert response.status_code == 422
+        before = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__AUDIO_TRANSCRIPTIONS}").count()
 
-        # Check that no log was created due to invalid JSON
+        # Get embeddings model id
+        response = client.get_without_permissions(f"/v1{ENDPOINT__MODELS}")
+        models = response.json()["data"]
+        model_id = [m for m in models if m["type"] == ModelType.AUTOMATIC_SPEECH_RECOGNITION][0]["id"]
+
+        file_path = "app/tests/assets/audio.mp3"
+        with open(file_path, "rb") as file:
+            files = {"file": (os.path.basename(file_path), file, "audio/mpeg")}
+            data = {"model": model_id, "language": "fr", "response_format": "json", "temperature": 0}
+            response = client.post_without_permissions(f"/v1{ENDPOINT__AUDIO_TRANSCRIPTIONS}", files=files, data=data)
+
+        assert response.status_code == 200
+
+        try:
+            for _ in response.iter_lines():
+                pass
+        except Exception:
+            pass
+        finally:
+            response.close()
+
+        # Allow some time for the asynchronous logging to complete.
         time.sleep(0.5)
-        current = db_session.query(Usage).count()
-        assert current == previous
+
+        after = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__AUDIO_TRANSCRIPTIONS}").count()
+        assert after - before > 0
+        log = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__AUDIO_TRANSCRIPTIONS}").order_by(Usage.id.desc()).first()
+        assert isinstance(log.datetime, datetime)
+        assert log.method == "POST"
+        assert log.model is not None
+
+    def test_classification(self, client, db_session):
+        """Test logging of a text classification request using log_usage decorator."""
+
+        before = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__RERANK}").count()
+
+        # Get a test classification model id.
+        response = client.get_without_permissions(f"/v1{ENDPOINT__MODELS}")
+        models = response.json()["data"]
+        model_id = [m for m in models if m["type"] == ModelType.TEXT_CLASSIFICATION][0]["id"]
+
+        params = {"model": model_id, "prompt": "Sort these sentences by relevance.", "input": ["Sentence 1", "Sentence 2", "Sentence 3"]}
+        response = client.post_without_permissions(url=f"/v1{ENDPOINT__RERANK}", json=params)
+        assert response.status_code == 200, response.text
+
+        try:
+            for _ in response.iter_lines():
+                pass
+        except Exception:
+            pass
+        finally:
+            response.close()
+
+        # Allow time for asynchronous logging to complete.
+        import time
+
+        time.sleep(0.5)
+
+        after = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__RERANK}").count()
+        assert after - before == 1
+
+        log = db_session.query(Usage).filter_by(endpoint=f"/v1{ENDPOINT__RERANK}").order_by(Usage.id.desc()).first()
+        assert isinstance(log.datetime, datetime)
+        assert log.method == "POST"
+        assert not log.prompt_tokens
+        assert log.completion_tokens is None
+        assert log.total_tokens is None
+        assert log.duration > 0
+        assert log.model is not None
